Move "NetIdHelper" to tests where its used
